Remove ROS dependencies
Remove hard dependencies on ROS Python2 packages like rospy, geometry_msgs etc. In the future, use loose coupling with ROS (like roslibpy) instead, which does not require the import of Python2 packages.

Replace all calls to the rospy logging system with Python's built-in logging module.

Refactor all usages of geometry_msgs etc. in helper.py to helper_deprecated.py, which is only used in the pycram_bullet_world_demo and local_transformer.py. Both, as well as IK.py, will have to be refactored at a later date, but at least the rest of PyCRAM does not require Python2 anymore.
