+++ demos/pycram_pr2_bullet_world_demo/donbot_process_modules.py
@@ -1,3 +1,4 @@
+import pycram.helper_deprecated as helper_deprecated
 from pycram.robot_description import InitializedRobotDescription as robot_description
 from pycram.process_module import ProcessModule
 from pycram.process_modules import ProcessModules, _apply_ik
@@ -135,8 +136,8 @@ class DonbotMoveHead(ProcessModule):
                     local_transformer.projection_namespace else \
                     solutions['target']
                 pose_in_neck_base = local_transformer.tf_transform(neck_base_frame, target)
-            elif helper.is_list_pose(solutions['target']) or helper.is_list_position(solutions['target']):
-                pose = helper.ensure_pose(solutions['target'])
+            elif helper_deprecated.is_list_pose(solutions['target']) or helper_deprecated.is_list_position(solutions['target']):
+                pose = helper_deprecated.ensure_pose(solutions['target'])
                 pose_in_neck_base = local_transformer.tf_pose_transform(local_transformer.map_frame, neck_base_frame, pose)
 
             vector = pose_in_neck_base[0]
